Change default `sim` feature to explicit `hw`
Previously the `sim` feature would determine if the enclaves would be linked
against the simulated vs hardware implementation of SGX. This has been
changed to a feature named `hw`. The change was made to more
consistently work with the `test_enclave`.  The cargo test flags of
`--no-default-features` only seemed to apply to the main crates of the
workspace and *not* the dev-dependencies.  This created an asymmetry
where the `test_enclave` would need to have an additive feature to
enable HW linking while the main crates had a subtractive feature to do
the same.