Convert to `sim` feature for SGX crates
Previously the SGX crates defaulted to simulated builds and used a
feature, `hw` to do hardware builds.  Now hardware builds are the
default and users need to use the `sim` feature to do sim builds.  The
default to hardware builds was chosen as it is less likely someone
accidentally builds and runs simulation intending for hardware builds.