Move `UnsealedData` into `mc-sgx-tservice` crate.
Previously the `UnsealedData` was in the `mc-sgx-tservice-types` crate.
There is little reason for `UnsealedData` to exist outside of an enclave
as the intent of sealing is to encrypt the data to store outside of an
enclave.  Encrypting data that existed, in plaintext, outside of an
enclave provides little to no benefit.

Also renamed `UnsealedData` -> `Unsealed` and `SealedData` -> `Sealed`.