Rework `mc-sgx-tservice::Report` trait
Previously the `mc-sgx-tservice::Report` trait had method
implementations was explicit to `mc-sgx-core-types::Report` and was
lacking an impl block. Now the `mc-sgx-tservice::Report` trait is
method signatures and there is an impl block for of the
`mc-sgx-tservice::Report` for the `mc-sgx-core-types::Report` type.

This also renamed the `mc-sgx-tservice::Report` trait methods:
- `create()` -> `new()`
- `verify_report()` -> `verify()`