Only support `sim` in test_enclave
Previously the `test_enclave` had a `sim` feature.  The `sim` feature
was used to build the `test_enclave` in a sim environment.  Cargo
currently doesn't support feature propagation to _only_ dev
dependencies. Always propagating `sim` to the `test_enclave` resulted in
failures to package `mc-sgx-urts` as `mc-sgx-urts` was now looking for a
published version of `test_enclave`.  Now `test_enclave` will only build
for the simulation SGX SDK, there is no feature flag to add or remove
this.