Use stable for rust builds
Previously only a nightly version of rust was used for CI and locally,
now the stable version is used. With nightly and beta used on CI for
build and test

The nighly is also used for the no alloc CI builds as these builds
require the `-Z` flag which is nightly only.