Add wrapper functions around nom's le_16 and le_32
The nom implementations of `le_16` and `le_32` use a generic error. In
order to use this with `expect()` one needs to specify the error type in
the function call with the turbofish operator. The wrapper functions
specify this error type and call `expect()` leaving the call sites of
`le_16` and `le_32` concise.