Fix copyright.yaml workflow
Previously the copyright.yaml workflow was using a quoted glob
expression for the files. Now the copyright.yaml workflow uses an
unquoted globstar expression.

The single quotes caused a failure in `action-update-license-year`:

    Found no files matching the path "'*.rs'\n'*.proto'"

Even with a globstar one would get:

    Found no files matching the path "'**/*.rs'\n'**/*.proto'"

Once the globstar and the quotes are dropped the action found the files.