Decode TCB values in PCK leaf certificate
Add the logic to decode TCB data off of the PCK leaf certificate. The
TCB values specific to an enclave are stored in the PCK leaf certificate
as OID extensions. Most importantly this contains the FMSPC value which
can be used to query for the advisories associated with these TCB
values at
<https://api.trustedservices.intel.com/sgx/certification/v4/tcb?fmspc={}>