Remove mounting .cargo/credentials
The binding of credentials was meant to make it easier to run
`cargo-release`. However if a user hasn't authenticated with crates.io
then this results in an error to create the container, which can be hard
to debug. See https://github.com/devcontainers/spec/issues/132 for a
potential way to support this in the future.