Fix tstdc crates calling out 0.7.0 dependencies
In the last release the tstdc crates didn't get all of the dependencies
updated correctly to '0.7.1' and were still calling out '0.7.0'. These
are now referencing the '0.7.1' crates.

Since we currently don't use the tstdc crates we will wait to publish
the fix until the next release needing overall changes from the SGX
crates.