Fix logic to update README of `mc-sgx-core-build`
When using `cargo release` one crate in a repository has settings on
how to update the CHANGELOG. Since this crate has the Cargo.toml entry
`[package.metadata.release]` it will not look at the entry in the root
`Cargo.toml`. The common settings of updating the README must be
repeated in this crate's `[package.metadata.release]`.