Update logic for new bitflags behavior
The bitflags crate was updated to allow for unnamed flags. This change
resulted in the formatting of `Attributes` to change slightly. As part
of this upgrade the formatting of `Attributes` has been improved to show
both the known flags as well as bitflag values that don't pertain to a
known flag.